Revert changes in transforms.
