+++ mart/transforms/vision/objdet/extended.py
@@ -6,11 +6,18 @@
 
 import logging
 import os
+from typing import Dict, Optional, Tuple
 
 import numpy as np
 import torch
 from PIL import Image, ImageOps
+from torch import Tensor
+from torchvision.transforms import functional as F
 from torchvision.transforms import transforms as T
+
+# FIXME: We really shouldn't be importing private functions
+from .torchvision_ref import ConvertCocoPolysToMask as ConvertCocoPolysToMask_
+from .torchvision_ref import _flip_coco_person_keypoints
 
 logger = logging.getLogger(__name__)
 
@@ -21,6 +28,9 @@ __all__ = [
     "SplitLambda",
     "LoadPerturbableMask",
     "LoadCoords",
+    "ConvertInstanceSegmentationToPerturbable",
+    "RandomHorizontalFlip",
+    "ConvertCocoPolysToMask",
 ]
 
 
@@ -103,6 +113,17 @@ class SplitLambda(Lambda):
         return tensor
 
 
+class ConvertInstanceSegmentationToPerturbable(ExTransform):
+    """Merge all instance masks and reverse."""
+
+    def __call__(self, image, target):
+        perturbable_mask = torch.sum(target["masks"], dim=0) == 0
+        # Convert to float to be differentiable.
+        target["perturbable_mask"] = perturbable_mask.float()
+
+        return image, target
+
+
 class LoadPerturbableMask(ExTransform):
     """Load perturbable masks and add to target."""
 
@@ -137,3 +158,56 @@ class LoadCoords(ExTransform):
         # Convert to float to be differentiable.
         target["coords"] = coords
         return image, target
+
+
+class RandomHorizontalFlip(T.RandomHorizontalFlip, ExTransform):
+    """Flip the image and annotations including boxes, masks, keypoints and the
+    perturable_masks."""
+
+    @staticmethod
+    def flip_boxes(image, target):
+        width, _ = F.get_image_size(image)
+        target["boxes"][:, [0, 2]] = width - target["boxes"][:, [2, 0]]
+        return image, target
+
+    @staticmethod
+    def flip_masks(image, target):
+        target["masks"] = target["masks"].flip(-1)
+        return image, target
+
+    @staticmethod
+    def flip_keypoints(image, target):
+        width, _ = F.get_image_size(image)
+        keypoints = target["keypoints"]
+        keypoints = _flip_coco_person_keypoints(keypoints, width)
+        target["keypoints"] = keypoints
+        return image, target
+
+    @staticmethod
+    def flip_perturable_masks(image, target):
+        target["masks"] = target["masks"].flip(-1)
+        return image, target
+
+    @staticmethod
+    def flip_perturbable_mask(image, target):
+        target["perturbable_mask"] = target["perturbable_mask"].flip(-1)
+        return image, target
+
+    def forward(
+        self, image: Tensor, target: Optional[Dict[str, Tensor]] = None
+    ) -> Tuple[Tensor, Optional[Dict[str, Tensor]]]:
+        if torch.rand(1) < self.p:
+            image = F.hflip(image)
+            if target is not None:
+                image, target = self.flip_boxes(image, target)
+                if "masks" in target:
+                    image, target = self.flip_masks(image, target)
+                if "keypoints" in target:
+                    image, target = self.flip_keypoints(image, target)
+                if "perturbable_mask" in target:
+                    image, target = self.flip_perturable_masks(image, target)
+        return image, target
+
+
+class ConvertCocoPolysToMask(ConvertCocoPolysToMask_, ExTransform):
+    pass
